# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "IT asset management API. Most routes require a bearer token from /auth/login.",
        "title": "Inventory Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "\u003cyour-app\u003e.up.railway.app",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate and receive a JWT",
                "parameters": [
                    {
                        "description": "Username and password",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getAssetList": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets"
                ],
                "summary": "List assets, paged and filtered",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Free-text search on code and name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "purchase_date",
                        "description": "Column to sort by",
                        "name": "sortField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "descend",
                        "description": "ascend or descend",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AssetListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AssetListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AssetResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.AssetResponse": {
            "type": "object",
            "properties": {
                "assetCode": {
                    "type": "string"
                },
                "assetName": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "createDt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "purchaseDate": {
                    "type": "string"
                },
                "serialNumber": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updateDt": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "models.LoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by the token returned by /auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}