# handler/auth.go
package handler

import (
	"context"
	"database/sql"
	"inventory-tracker/models"
	"inventory-tracker/utils"
	"net/http"

	"github.com/gin-gonic/gin"
)

func (h *Handler) Register(c *gin.Context) {
	var req models.RegisterReq
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Check if username already exists
	var exists bool
	err := h.DB.QueryRow(context.Background(),
		"SELECT EXISTS(SELECT 1 FROM users WHERE username=$1)",
		req.Username,
	).Scan(&exists)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	if exists {
		c.JSON(http.StatusConflict, gin.H{"error": "Username already exists"})
		return
	}

	// Hash password
	hashedPassword, err := utils.HashPassword(req.Password)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to hash password"})
		return
	}

	// Insert user
	var user models.Users
	err = h.DB.QueryRow(context.Background(),
		`INSERT INTO users (name, category_group, username, password) 
         VALUES ($1, $2, $3, $4) 
         RETURNING id, name, category_group, username`,
		req.Name, req.CategoryGroup, req.Username, hashedPassword,
	).Scan(&user.Id, &user.Name, &user.CategoryGroup, &user.Username)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create user"})
		return
	}

	// Generate token
	token, err := utils.GenerateToken(user.Id, user.CategoryGroup)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate token"})
		return
	}

	c.JSON(http.StatusCreated, gin.H{
		"message": "User registered successfully",
		"user":    user,
		"token":   token,
	})
}

// @Summary  Authenticate and receive a JWT
// @Tags     auth
// @Accept   json
// @Produce  json
// @Param    credentials body models.LoginReq true "Username and password"
// @Success  200 {object} map[string]string
// @Failure  400 {object} map[string]string
// @Failure  401 {object} map[string]string
// @Router   /auth/login [post]
func (h *Handler) Login(c *gin.Context) {
	var req models.LoginReq
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	var user models.Users
	err := h.DB.QueryRow(context.Background(),
		"SELECT id, name, category_group, username, password FROM users WHERE username=$1",
		req.Username,
	).Scan(&user.Id, &user.Name, &user.CategoryGroup, &user.Username, &user.Password)

	if err == sql.ErrNoRows {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid credentials"})
		return
	} else if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Database error"})
		return
	}

	// Check password
	if !utils.CheckPasswordHash(req.Password, user.Password) {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid credentials"})
		return
	}

	// Generate token
	token, err := utils.GenerateToken(user.Id, user.CategoryGroup)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate token"})
		return
	}

	// Don't send password back
	user.Password = ""

	c.JSON(http.StatusOK, gin.H{
		"token": token,
		"user":  user,
	})
}

func (h *Handler) GetCurrentUser(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "User not authenticated"})
		return
	}

	var user models.Users
	err := h.DB.QueryRow(context.Background(),
		"SELECT id, name, category_group, username, created_at FROM users WHERE id=$1",
		userID,
	).Scan(&user.Id, &user.Name, &user.CategoryGroup, &user.Username)

	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
		return
	}

	c.JSON(http.StatusOK, user)
}